+++ projects/net.splitcells.gel.core/src/main/java/net/splitcells/gel/data/lookup/LookupColumn.java
@@ -24,9 +24,7 @@ import java.util.List;
 import java.util.ListIterator;
 import java.util.Optional;
 import java.util.function.Predicate;
-import java.util.stream.Collectors;
 
-import net.splitcells.dem.utils.NotImplementedYet;
 import net.splitcells.dem.data.set.list.Lists;
 import net.splitcells.gel.data.table.Line;
 import net.splitcells.gel.data.table.Table;
@@ -195,7 +193,7 @@ public class LookupColumn<T> implements Column<T> {
 
     @Override
     public net.splitcells.dem.data.set.list.List<T> values() {
-        return table.linesStream()
+        return table.unorderedLinesStream()
                 .map(e -> e.value(attribute))//
                 .collect(Lists.toList());
     }