+++ projects/net.splitcells.gel.core/src/main/java/net/splitcells/gel/data/table/Table.java
@@ -100,18 +100,18 @@ public interface Table extends Discoverable, Domable, Identifiable {
      * This is caused by the fact, that this method does not need a copy all {@link Line} references,
      * before constructing this {@link Stream}.</p>
      *
-     * <p>Also, it is expected, that generally {@link #linesStream()} is the fastest method,
+     * <p>Also, it is expected, that generally {@link #unorderedLinesStream()} is the fastest method,
      * in order to iterate or list all {@link Line}.
      * Especially, it is expected to be generally faster, than {@link #rawLines()}.
      * Only in select cases {@link #rawLines()} is expected to be faster,
      * where knowledge of the underlining {@link Table} implementation is present.
-     * The reason for this, is the fact, that it is faster to construct {@link #linesStream()} via {@link #rawLines()},
+     * The reason for this, is the fact, that it is faster to construct {@link #unorderedLinesStream()} via {@link #rawLines()},
      * than the other way around.</p>
      * <p>There is no guarantee regarding the ordering.</p>
      *
      * @return An ordered {@linl Stream} of {@łink #lines}.
      */
-    default Stream<Line> linesStream() {
+    default Stream<Line> unorderedLinesStream() {
         return unorderedLines().stream();
     }